lsp-did-open: send text from kakoune to kak-lsp
kak-lsp is used to read a file's content from a filesystem on `didOpen`
before sending it to a language server, and this is something that does
not work in case we are dealing with newly created (unsaved) file. In
order to fix that problem, we need to send a buffer context from kakoune
directly to kak-lsp, similarly to `didChange` message.

Closes #156