Allow project roots to be specified by environment
Allow per language marker files to be overridden by directly specifying
the project root for a given buffer. When searching for the project root
for a buffer first see if there are any environment variables prefixed
by `KAK_LSP_PROJECT_ROOT_%LANG%`. If so then iterate over all matching
variable values checking if they are a prefix of the path to the current
buffer. If so the first matching prefix is selected as the project root,
else return the buffers path. If there are no matching environment
variables then proceed as before.