Don't display an error message for a CONTENT_MODIFIED error.
According to the lsp specification [1], "if clients receive a
ClientModified error, it generally should not now it in the UI for
the end-user. Clients can resend the request if appropriate."

Note that this patch doesn't resend the request.

Since there's only one language-server error that we recognize, this
patch hard-codes it. There are currently two lsp-specific error codes,
so it doesn't make sense to pull in any dependencies for their
definition. If it ever does, the "lsp-server" crate exports the error
codes.

[1] https://microsoft.github.io/language-server-protocol/specification