Use textEdit completions that replace token left of cursor
This is a targeted fix for the cases where Kakoune insertion would do
exactly the same as the text edit.
If the language server sends simple textEdit objects that replace
precisely the token left of the cursor, then we can just pass them to
Kakoune as normal completions. This is strictly better than using the
completion's label, because Kakoune does exactly what the text edit
is supposed to do by default.

See below for an example that is not fixed by this commit. I believe
it makes sense to have this partial support until #40 is properly
implemented, because it seems safe and fixes what is arguably the most
common completion scenario: completing the suffix of a token.

Reproduce with this example C file using ccls. Cursor is |.

	void foobar(){}
	void main(){
		fo|	// Fixed:	Completion inserts foobar()
		fo|ba	// Not fixed:	Completion inserts foobar() -> void
	}

I thought about also changing the behavior for the "Not fixed" case,
where we could also just pass on the completion.
The result would be that Kakoune replaces "fo" with "foobar()",
ignoring the existing suffix, so we'd end up with "foobar()ba".
This is better than the labels here, but I'm not sure if it's
a good idea to knowingly leak this error.

Improves on #275