Fill out all client capabilities manually.
This actually adds a few things we previously weren't asking for, so
it might actually give us some free new features for language servers
that were checking that. It also makes for a good place to find more
things to work on! There's still a lot that can be improved.