Avoid noisy errors when default language servers fail to start
Commit fe24812 ("Show an error popup when the language server failed to start")
was not really a good idea: users often run the default config, with language
servers configured for many languages. They usually don't have all of them installed.
When we are enabled globally with "lsp-enable" and the user edits a file
where the corresponding language server is not installed, they will get
annoying popups.

I'm really not sure what the best solution is.
We could cache the fact that the language server failed to start, and refuse
to restart it as a consequence. However, then we'd need to invalidate that
cache.

We might have to revert fe24812 in the end, however, first try to
use this simple heuristic: only report the error if a path contains
a slash. This is never the case for the default config, and
it helps in the motivating issue #423.