Show an error popup on invalid config
In https://discuss.kakoune.com/t/kak-lsp-for-julia-languageserver-jl/1389
user configuration was missing the required "roots" parameter.
The log file can easily diagnose that; this helps users who didn't
set up a log file yet.

There is a pathological case with this change when viewing the *debug*
buffer with an invalid while running an invalid kak-lsp.toml.
Starting kak-lsp will fail and log an error to the *debug* buffer, which
means that the buffer changed, and on the next keystroke we will
send another request with kak-lsp. This means that every keystroke
triggers this diagnostic. However, that's an edge case, and one
should just fix the config error.