Show an error popup when the language server failed to start
This is usually a misconfiguration - probably the wrong path.  Tell the user
immediately. I can't think of a situation where one would want to ignore
this error.

Add the executable's filename to errors that correspond to ENOENT and EPERM.
The error is shown in an ":info" popup and logged to the *debug* buffer.
Maybe we should use ":fail" for such errors in future, to match other errors
in Kakoune.

See #423.