Don't change Kakoune's directory when using diagnostics/goto
For buffers *diagnostics* and *goto* we used Kakoune's "make" and "grep"
filetypes respectively. This commit changes that by copying and adapting
relevant parts from those files.

This allows us to
1. customize the "grep-jump" command, to respect a new %opt{lsp_project_root}.
   Jumps are made relative to this option, so they work even when Kakoune's
   working directory is different from the project root.
2. "lsp-next-location" to use "lsp-jump" instead of "grep-jump" [1].
   The advantage of "make-jump" is that it shows the error message in the
   prompt line. Teach "lsp-jump" the same trick by adding a buffer-local
   alias, overriding plain "lsp-jump".

   This alias is now also used by "lsp-next-location" which therefore also
   shows the error message - like "make-next-error", and unlike "grep-jump"
   which was used before.
3. Set a single *lsp-goto* filetype for *goto*, *diagnostics* and similar
   buffers.  We could also use a separate filetype for *diagnostics* because
   it uses slightly different highlighting and jumping, but I don't know if
   it's worth exposing this to the user.

This also requires us to vendor the "make" highlighting logic, because we
want to modify it: our "lsp-next-location" uses "grep_current_line" and not
"make_current_error_line", so highlighting must do the same.

We still depend on some options defined by grep.kak; that could be changed
if ever needed.

Closes #406

[ja: tweak patch and commit message]