Send error message to editor when server command contains spaces
There were multiple reports of users incorrectly assuming that "command'
is a shell command that contains arguments, for example
https://discuss.kakoune.com/t/i-have-no-idea-if-kak-lsp-is-running/1093/15

If the server fails to start, and the command name contains a space,
send an error like this to the editor:

	Failed to start language server: No such file or directory \
	(os error 2): the-language-server some args

This message shows up in the *debug* buffer and in an :info message.
For some reason, the :info message only seems to work starting with
the second time the language server is started.