Fix parsing of nested dynamic initialization options
We support dynamic configuration via %opt{lsp_server_configuration}.
For example,

	set-option global lsp_server_configuration rust.features=["the-feature"]

will tell rls about our project's feature flag, enabling it to find functions that
are only defined when this feature is active:

	#[cfg(feature = "the-feature")]
	fn gated_function() {}

This works because we convert

	rust.features=["the-feature"]

to a nested object, which we send to the language server

	{"rust":{"features":["the-feature"]}}

Curiously, even VSCode uses the first notation in their settings.json.

rust-analyzer supports the same feature, but instead of using server
configuration, they use initialization options.  So the equivalent
to the above rls configuration would be

	set-option global lsp_server_initialization_options cargo.features=["the-feature"]

Prior to this commit this was not supported. We would blindly send
{"cargo.features":{"the-feature"}}.
Fix that by applying the same conversion as for server configuration.

Fixes #480