Send textDocument/didChange synchronously to prevent race with completion requests
lsp-did-change-and-then first sends textDocument/didChange
asynchronously, and then invokes its callback request.  If the
textDocument/didChange arrives after the callback request, then
language server would use a stale version of the document.  Try to
fix that by waiting until the first request has been sent.

Closes #267

This also helps with a use case that arose in #582: if two *-sync
commands run in a pre-write-hook, and the first one modifies the
buffer, the second one would execute asynchronously, if it uses
lsp-did-change-and-then. Now this will work:

	hook -group my-format global BufWritePre .*[.]go %{
		try %{ lsp-code-action-sync '^Organize Imports$' }
		lsp-formatting-sync
	}