Properly implement Semantic Token modifiers
Until now modifiers used to overwrite a token mapping, rather than
allowing to alter it.
With the new configuration, faces are now mapped to a combination of
both token plus an arbitrary set of modifiers, meaning that `variable`,
`variable declaration` and `function declaration` can be mapped to
three different faces.

Signed-off-by: Lucas Schwiderski <lucas@lschwiderski.de>