Introduce %opt{lsp_config} for dynamic server configuration
Prior to this commit, server-specific settings could be configured with either of:
- `[language.<filetype>.initialization_options]` in `kak-lsp.toml`
- `%opt{lsp_server_initialization_options}`
- `%opt{lsp_server_configuration}`

There are some problems with the latter two; see #500.

This commit attempts to create a simpler and more predictable interface
by allowing to set `[language.<filetype>.initialization_options]`
dynamically in the new option `%opt{lsp_config}`.

Unlike the old dynamic config options, this is a raw TOML string,
with exactly the same format as kak-lsp.toml. This makes sure we can
express all nested object structures that servers want, and don't
have to learn special rules.

For example:

	set-option global lsp_config %[
		[language.go.initialization_options]
		buildTags = "sometag"
		"formatting.gofumpt" = true
	]

This currently only supports
`[language.<filetype>.initialization_options]` but in future we can
add support for other fields from `kak-lsp.toml` (hence the "config"
in the name).  It would be cool if you could change every config
option dynamically, then there would be no more need to create a
`kak-lsp.toml`.

Initialization options are now computed by using the first among the
following options where the given language has a non-empty config.
1. `[language.<filetype>.initialization_options]` from `%opt{lsp_config}`
2. the deprecated `%opt{lsp_server_initialization_options}`
3. `[language.<filetype>.initialization_options]` from `kak-lsp.toml`

A previous iteration implementated a merge logic for the first two,
but that seems unnecessarily complex; there should be no good reason
to use the old config syntax.

Likewise, the options to `workspace/DidChangeConfiguration` are
computed by checking
- `[language.<filetype>.initialization_options]` from `%opt{lsp_config}`
- the deprecated `%opt{lsp_server_initialization_options}`
Here we could also use the static configuration as fallback, but
we don't need to since that's already sent during initialization.

So this means that with this commit,
`[language.<filetype>.initialization_options]` from `%opt{lsp_config}`
feeds into both `initialize` and `workspace/DidChangeConfiguration`,
instead of just the former.  This is a deliberate behavior change;
I can't think of a scenario where this is worse.

The name "initialization_options" seems a bit stale but it seems
alright. We could use "settings", which is what LSP uses for this
general thing.

Closes #500