Remove lsp-show-message in favor of individual functions for each log level
The old implementation ran "info", but without a client in the context,
which doesn't work. This means that we didn't report some error
messages at all, for example when we send wrong initialization_options
to gopls. This helps a lot when investigating issues like #500.

Since we receive these messages from the server, we don't really have
a meaningful client. Use %opt{toolsclient} mostly because it fits
my workflow. As fallback, use the client that ran "lsp-enable", even
though it might not be alive anymore.  Maybe we should send it to all
clients, or somehow remember the client with the most recent activity?

The message type parameter was an integer which can be convenient
(think `test $1 -gt 2`) but a little copying + more native Kakoune
feel seems better.

No need for deprecation since this appears to be unused according to
a quick GitHub search, plus the default implementation was broken.