Use custom faces for Markdown highlighting
There are too many variables involved in creating the final colours that
a user sees in their Kakoune instance. Trying to guess highlighting
colours would likely only work for a small part of the user base.

Instead we emit custom faces in the Markdown parsing. We don't actually
define these faces in Kakoune, though, so by default the text is not
highlighted, but readable.
Users and/or theme creators can then assign colours to those faces that
fit their colour scheme.