Fix check if text edit insertion will merge selections
When applying multiple text edits to the same document,
we first create a selection for each text edit
According to the spec, the text edits must never overlap.
However, after applying a text edit, adjacent selections might
overlap and are therefore merged by Kakoune. Our merged_selections()
computes which edits create such overlapping selections.  This commit
fixes a case where merged_selections() fails to recognize that two
selections will merge.

The root cause is in lsp_range_to_kakoune_utf_8_code_units()
where we have this logic

    // Exclusive->inclusive range.end conversion will make 0-length LSP range into the reversed
    // 2-length Kakoune range, but we want 1-length (the closest to 0 it can get in Kakoune ;-)).
    let end_byte = if insert {
        start_byte
    } else if end.character > 0 {
        // -1 because LSP ranges are exclusive, but Kakoune's are inclusive.
        end.character - 1
    }

It converts LSP ranges to Kakoune ranges, in an inconsistent way.
It converts:
- 0-length LSP range -> 1-length Kakoune range
- length n LSP range -> n-length Kakoune range (where n > 0)

If the current range is non-empty but the next range is empty (so
it's an insertion), then merged_selections() failed to recognize if
they merged, because of above inconsistency.

Fix that by detecting if the next range is empty in "second_is_insert",
and rename the other variables accordingly.

Fixes #521