Log the "roots" glob that found the project root
Commit d87447e ("Recognize VCS dirs as roots for C/C++ in the
default config") was ineffecitve on systems that had directories like
"~/.clangd".  This is because the "roots" config option is traversed
first-to-last.  If the roots parameter is [".clangd", ".git"], then
".git" will only be used if there is no parent directory that contains
".clangd".

We could switch the order, so that parent directories are only
searched if the current directory contains none of the roots.
However, that might cause problems, I'm not sure.

Add a log entry so that users can debug this scenario without
recompiling.