Fix check if text edit replacement will merge selections
Extend "remove_adjacent" to include non-empty replacements.  The attached
test case was extracted from a rust-analyzer example. It's not
minimized; the crucial fix here is that "lsp-replace-selection
'println'" is recognized to merge with its (right) neighbor selection.

Also change the logic for "remove_adjacent" from
(empty && adj1 || adj2) to (empty && (ajd1 || ajd2).
This is arguably more correct.

Because we now recognize more merged selections, we must take care of a
special Kakoune behavior: "|" commands that that output the same
text that was input, preserve all selection.  This messed with our
logic that would assume they would merge (and go away).  Luckily,
we can just drop such redundant text edits.  This causes harmless
changes to the existing test.

Part of #527