Fix text edits with leading newline by adding sentinel character
https://github.com/mawww/kakoune/issues/4373 means that we cannot
reliably insert text that starts with a newline, because it
will stomp other text edits' selections.
Work around this by adding a sentinel character, so that no insertion
actually starts with a newline character. Take a character from the
Unicode private use area, because we can assume that it doesn't occur
in text documents. This way we can simply delete all occurrences after
we're done.

Closes #527