Don't drop trailing newline in text edits
Kakoune's | command includes special handling of selections that do
not end in a newline: a single newline is removed from the output.
A text edit that replace text (that does not end in a newline) was
affected by this if it inserted text with a trailing newline.
Don't drop the newline in this case. This fixes a case where we would
fail to format this line using rust-analyzer's text edits:

	fn main() { foo("".to_string()); }