Create just one undo entry for text edits with client context
rust-analyzer likes to send elaborate sequences of text edits for
code actions. We used to apply them inside an "eval -buffer", which
means that they run without a Kakoune client in context.  This means
that the active client would notice each individual text edit and
create an undo entry for every single one.

Use "eval -buffer" only when necessary (when there is no client in
context), otherwise just open the buffer in a draft context so all
edits are collapsed.

This changes the behavior for #504, but it feels more consistent now.

As suggested in https://github.com/kak-lsp/kak-lsp/issues/527#issuecomment-922573440

This fixes the cases where the target buffer was displayed in
some client.  Otherwise, Kakoune would only reload the buffer when
displaying them in a client, so you'd always get only one undo
entry anyway.