Ignore invalid UTF-8 when computing locations in files
When computing jump locations, if the language server does not provide
us with a file's contents, we need to read it ourselves. Previously,
we returned None in get_file_contents(), and didn't even log the error.

I'm not aware of a portable method to detect character encoding.
Luckily, most encodings will be ASCII compatible, so we can just treat
them like UTF-8 but replace invalid code points with a replacement
character.

Don't change apply_text_edits_to_file() because there we actually care
about the file contents, since we write them back.  Users will get
an error message in logs if they run "lsp-rename" that would result
in renaming a file that is not opened as a buffer.

The test is not yet run in the CI since it requires clangd; last time
I tried, the clangd package on GitHub Actions' Ubuntu was too old IIRC.

Closes #535