Enable fuzzy search in code action menu
Kakoune's "menu" command spawns a UI element that allows to select
one of a list of options. Currently the UI supports Tab and Shift+Tab,
or <c-n> and <c-p> to cycle through the options.  Additionally, /
can be used to enter a filtering regex.

When there are lots of code actions, the filtering can be very
convenient.  However, fuzzy search among prompt completions seems
even better (and more familiar), so define a temporary command with
the appropriate completions, and run that instead of "menu".

Simplify a related test case that's too specific anyway.

Closes #539