Add test for code actions using rust-analyzer
I'm not sure if these tests are really useful in the long run.
They are fairly slow and test many things that are not something
that we would fix. We should also figure out a way to automatically
update servers.

Probably we should test on the protocol level instead.

However, they're easy to write and read, so they are nice as
documentation.