test: fix XDG dir setup
XDG_RUNTIME_DIR must exist with the 700 mode.  Also, XDG_DATA_DIR is
not a thing; it should be XDG_DATA_HOME. Since this is not needed,
remove it. We only need config and runtime for both kak-lsp/kak.
In future we might also need to isolate cache/state/data directories
for language servers.

Only set the variables if they are defined in the environment
(otherwise we already use safe fallbacks) to avoid needlessly creating
xdg_runtime_dir.