test: do not run kak as child of tmux, to make servers exit
For some reason, when launching kak like "tmux new-session" and
quitting it, our KakEnd hook does not manage to send the "stop"
request to the kak-lsp server.  As a result, we also don't send LSP
exit notification so we also don't stop the language server.

Work around this by spawning an interactive shell to launch kak.
This allows us to cleanly exit both kak and tmux.  Then we expect
that no server is running, so we can clean up our scratch directory.