Do not request code actions by default
With the default idle_timeout of 50ms we send quite a lot of requests
when idle (#429).
When making nontrivial changes that break the build of the kak-lsp
project, the code action requests seem to keep rust-analyzer maxing
out 1-2 CPUs for a while, even causing some lag in the editor UI
(at least with a debug build of kak-lsp).

The lightbulb is not worth the risk of breaking basic functionality
(like go-to-definition), so disable it by default. Add option
lsp_auto_show_code_actions to enable sending requests.