Add the ability to goto/get hover info of the next/prev symbols in a buffer
You are deep within a huge function. You have forgotten the name of
the function you are in or want to know what its parameters are. This
patch provides the Kakoune command lsp-hover-previous-function,
which shows a modal with the hover info of the previous function.
In the modal, press g to actually jump to the function definition.
without your cursor actually going to the previous function.
Any other keys dismiss the modal.

Similarly, new command lsp-{previous,next}-function will directly go
to the previous/next function.

Same for any kind of symbol exposed by LSP.

Patch-tweaked-by: Johannes Altmanninger <aclopte@gmail.com>

Closes #557