Drop progress report notifications that arrive too quickly
On startup, rust-analyzer can send one progress notification per
millisecond. This timespan is well beyond human perception, and the
number of messages can can slow us down because we spawn a kak process
whenever we relay a notification to lsp-handle-progress.

Progress reports are volatile report messages, which are always guarded
by begin/end messages, so it should be fine to just drop some of them
(the end message will make sure that we clear the progress value).
Drop every report that arrives within 50ms of the last accepted
report. This is a simple solution that should be good enough for
most cases. In future we can expose all progress reports to clients,
maybe by sending them in bulk.

In case we get two reports in quick succession followed by a long
delay, we will drop the second one, which is a potential problem.
However, that's unlikely to be a real issue. The main point
of these messages is to convey to the user that *something*
is happening, the details are secondary.