lsp-document-symbol: use selectionRange, to jump to definition start
Unlike the deprecated SymbolInformation type, DocumentSymbol have
this extra property.  This is a subrange of the entire range, but
typically does not include doc comments, so it seems more useful.

Tell servers that we support DocumentSymbol, so rust-analyzer actually
sends it, and use that range for lsp-document-symbol.

See the discussion in https://github.com/kak-lsp/kak-lsp/pull/557#discussion_r748691580