lsp-document-symbol: handle children of DocumentSymbol
As reported by Sidharth Kshatriya, the parent commit that requested
hierarchical DocumentSymbol data (instead of the legacy flat
SymbolInformation) broke in scenarios where we get nested symbols.
For example, on Kakoune's src/client.cc, lsp-document-symbols only
shows the namespace.

This is because we simply did not handle children; do that.