Functionality to allow hover information to be show in a new kak client
Sometimes the hover information sent to kakoune by kak-lsp is either
very long or has rich markdown (or both). The hover info is shown in an
info box which can be very difficult to read.

Add a new command `lsp-hover-in-hover-client` which will open the hover
information in a new client. You also get the benefit of syntax
highlighting as the codeblocks are enclosed within three ticks
(with the language name) and kak understands that.

The user can continue issuing `lsp-hover-in-hover-client` commands
and they will appear in the `hoverclient`. The focus will not
shift from the window in which the user was typing in.

This feature can be very visually appealing if you use kakoune
with something like tmux however it will work properly even without it.

If a `hoverclient` does not exist it will be created via `new`.

**Try this out**

If you are using rust-analyzer then place your cursor over any
of the characters of `fn` or any of the characters of `let` and
issue the above command (keyboard shortcut is lsp H).

You will see great documentation for `fn` / `let`.

**Why not just use goto definition?**

Many languages have the documentation for a function above it.
But you cannot goto definition in all situations. For instance, you
cannot goto definition on the `let` statement in Rust.

This feature can be especially useful for ocamllsp to be able to
see long signatures. In my testing other LSPs that give you
a lot of hover info become more usable.