Introduce lsp-code-action, to run a specific code action, bypassing the menu
Sometimes, I want to repeatedly execute a single code actions (in
different places). Our lsp-code-actions (plural) invokes a menu,
so I need to search for the right one every time.  We do allow to
override lsp-perform-code-actions to filter code actions, or run
select ones directly, but this is not a very nice UI because the
override is hard to undo.

Instead, add a new command lsp-code-action (singular) that takes a
regex, and executes the matching code action (if unique).  This means
that users can add bindings for something like ":lsp-code-action
Organize.Imports". It's still async, so it can't really be scripted
but it's a start.

The regex syntax is from the https://docs.rs/regex/1.3.9/regex/ It's
sufficiently similar to Kakoune regex, and it's mostly used letters,
so it doesn't seem necessary to document this for now.

Close #566