Honor extra_word_chars option when finding start of token to complete
lsp_completion_fragment_start uses \w+ to backtrack to the start of the
current token. This fails for some languages, notably Lisps that allow
other characters in identifiers besides alphanumeric and underscore.
Kakoune already knows the extra_word_chars option, which modify the
behavior of w, b and e.
Include extra_word_chars in the regex. This fixes the following
example for the racket language:

First install the racket, and then use this command to install their
language server:

	raco pkg install racket-langserver

Then create info.rkt with this content

	#lang slideshow
	(define a###b 123)
	(define b a#

and (since Kakoune doesn't ship with a racket file type yet) run

	:set window filetype racket

observe that completion will insert a#a###b.
This command will fix that, so we get a###b.

	:set window extra_word_chars '_' '-' '#'

See the discussion in #568