Refactor: simplify next/prev symbol detection
When exceeds() was introduced, it was called from two different
functions, but that is no longer true since 10ee06d (Deduplicate
DocumentSymbol/SymbolInformation handling, 2021-11-17).

We recently added a second call to exceeds(), to fix the
previous-symbol case. This uses a lot of vertical space and looks ugly.

Inline exceeds; this is actually simpler and allows us to separate the
"is this a symbol we care about?" logic from the position checks.

While at it, reword a comment, to try to explain why the
previous-symbol fix was necessary.  Also reduce the scope of the
values from the recursive call.

See the discussion in #569