Teach lsp-auto-hover-enable to show buffer in given client
Add support for the new lsp-hover-buffer in auto hover.
This allows a command like "lsp-auto-hover-enable %opt{docsclient}"
to automatically show hover info for the current cursor position
in the named client.  The scratch buffer avoids many of
the longstanding issues of the info box (size limits, not
searchable/editable, obstructs rest of the buffer).

Closes #257
Closes #570