Allow to show hover information in a scratch buffer
Sometimes the hover information sent to Kakoune by kak-lsp is either
very long or has rich markdown (or both). Hover info is shown in an
info box which can be very difficult to read.

Add a new command "lsp-hover-buffer" to render hover information in
a scratch buffer called *hover*. If the server sends Markdown, set
the markdown filetype, which adds syntax highlighting in code blocks.

lsp-hover-buffer takes an optional client argument that defaults to
%opt{docsclient}. If given, show the *hover* buffer in that client
but do not switch focus. If the named client does not exist,
create it with ":new".
If no client name (and %opt{doscclient} is empty), use the current
client.  This is consistent with Kakounes {git,doc,man,grep,make}.kak
and will work well in any kind of windowing environment.
Most of those do not mark the hover buffer readonly, so we do not
either.

This feature can be very visually appealing if you use Kakoune with
something like tmux.

**Try this out**

If you are using rust-analyzer then place your cursor over any of
the characters of "fn" or any of the characters of "let" and issue
the above command (keyboard shortcut is lsp H).

You will see great documentation for "fn" / "let".

**Why not just use goto definition?**

Many languages have the documentation for a function above it.
But you cannot goto definition in all situations. For instance,
you cannot goto definition on the "let" statement in Rust.

This feature can be especially useful for ocamllsp to be able to
see long signatures. In my testing other LSPs that give you a lot of
hover info become more usable.

Patch-by: Sidharth Kshatriya <sid.kshatriya@gmail.com>

Closes #75 (the upstream issue is already fixed)
Closes #387
Closes #409
Closes #514
Closes #564