Use width of error diagnostics as default gutter width
As reported by 0phoff, setting lsp_diagnostic_line_error_sign and
friends to strings with a display width > 1 will make the gutter
shift width to 1 (if there are no diagnostics) and back to width >
1 if there are some.

So far we assumed that gutter signs are always one character wide,
but it's valid to use more characters.
Eventually we want to compute the maximum of all widths, but for now
let's go with an 80% solution and just use the width of the error
symbol as default.

This means that all other symbols must have an equal (or lesser)
width, otherwise there will be more jumping.