Apply additionalTextEdits after completion is accepted
One of the uses of completionItem/resolve is to compute edits to
be applied when the completion is accepted. Support this.

For example, this makes kak-lsp insert
import statements from rust-analyzer, see
https://github.com/kak-lsp/kak-lsp/issues/40#issuecomment-991939958

This adds some noisy indentation changes to the detection of completion
offsets / simple text edits but "git diff -w" takes care of that.