Use UTF-8 offsets if supported by the server (clangd extension)
Today, kak-lsp defaults to UTF-16 offsets (per the spec) but users
can override that behavior per language.

To get rid of the need to configure this, implement the clangd protocol
extension [1] to negotiate the offset encoding. Tell servers that
we prefer UTF-8, and use that if they support it.

clangd supports this extension but gopls and pyls don't, so keep
their explicit offset_encoding override.

[1]: https://clangd.llvm.org/extensions.html#utf-8-offsets

Closes #485
Closes #189