Remove workaround for servers that send plaintext labeled as Markdown
This was only relevant for pyls; since the parent commit switched
to pylsp, there is no more need for this workaround. Remove it since,
this is really easy to fix for servers.

Keep the config key
workaround_server_sends_plaintext_labeled_as_markdown around as nop. We
encourage users to migrate, but it doesn't make sense to introduce
a gratuitous breaking change that'd also affect other languages.