Add request option to write raw LSP response to fifo
When "write_response_to_fifo" is set to true, kak-lsp will print the language
server's JSON response to the given fifo, instead of passing it to the editor.

This allows to write a script like the one below that prints the list of
all symbols in a given file.

We could hide fifo creation, but I don't think that helps really, because the
user still needs to construct a sizeable TOML input.  Since most LSP requests
expect an interactive editor context, the next patch will introduce a Kakoune
function that uses write_response_to_fifo", which hides most of the ugliness.

One thing to note is that positions passed via "[params]" are Kakoune
coordinates (1-based), but LSP responses are 0-based.  In future, we could
offer a different output format with preprocessed values in Kakoune coordinates
(perhaps as TOML, or plain text location lists).

See #580

Here is the example script I used to test.

	#!/bin/sh
	kak_session=session # Only works if a server for this session is up.
	kak_buffile=$PWD/src/main.rs
	kak_opt_filetype=rust
	kak_timestamp=1

	tmp=$(mktemp -d)
	pipe=${tmp}/fifo
	mkfifo ${pipe}

	printf "
	session  = \"${kak_session}\"
	buffile  = \"${kak_buffile}\"
	filetype = \"${kak_opt_filetype}\"
	version = \"${kak_timestamp}\"
	fifo     = \"${pipe}\"
	method   = \"textDocument/documentSymbol\"
	write_response_to_fifo = true
	[params]
	"  | kak-lsp -s ${kak_session} --request &

	cat ${pipe}
	rm -r ${tmp}