Add lsp-code-action-sync to synchronously run a code action
lsp-code-action can't be used safely in a BufWritePre hook because
it is asynchronous, which means that it races with the write.  Add a
synchronous variant (similar to lsp-formatting-sync) to support that.

Whereas the asynchronous variant shows errors in the info box
when there is no code action (or there are multiple matching ones),
lsp-code-action-sync shows them via "fail", because this is easy to
suppress with "try". Maybe we should silence these errors instead?

Since code actions are meant for interactive use,
they use two requests:
1. textDocument/codeAction asks the server for code actions
2. Once we have selected a code action to perform, we ask the kak-lsp
   server to do that. This happens via lsp-apply-workspace-edit
   or lsp-execute-command (which needs a roundtrip to the language
   server).  This means that we also need to add synchronous variants
   of lsp-apply-workspace-edit and lsp-execute-command; do that.

A synchronous request means that we need to return a Kakoune command
in every code path, so remove an early return that may prevent us
from doing so.

Closes #582