Cache workspace/configuration response instead of asking editor
Our "lsp-*-sync" commands block until kak-lsp returns a response.

When kak-lsp sends a textDocument/formatting request, texlab[1]
asks for workspace/configuration before answering.
For workspace/configuration we ask the Kakoune session for the
current value of lsp_config (via lsp-get-config). Since Kakoune is
blocked, this is a deadlock.

Fix this by caching lsp_config in kak-lsp.  The editor alreay sends
it to kak-lsp in the workspace/didChangeConfiguration hook, so we
can just tap that. We still keep lsp-get-config for initialization.

Fixes the issue with lsp-formatting-sync mentioned in
https://github.com/kak-lsp/kak-lsp/pull/573#issuecomment-989350761

It currently doesn't make sense to have different values of lsp_config
for different buffers, because we only support one language server per
filetype, so it's okay to use a global for lsp_config.  When multiple
languages are present in lsp_config, it's best to set it in global
scope; when using buffer scope, kak-lsp might mix up the configs
after this change.

[1]: https://github.com/latex-lsp/texlab/issues/533