Declare that *-sync commands require a running server
If no server is running (for example right after starting the editor,
without running lsp-enable), then commands like lsp-formatting-sync
would block forever. This happens because server initialization needs
lsp-get-config, which doesn't work if the editor thread is blocked.

If the kak-lsp server is running, but the language server is not
(for example after the timeout), then a lsp-formatting-sync will
be canceled ("cancelling blocking request"). Luckily, this scenario
shouldn't happen very often, because the default timeout is quite high,
and we restart language servers whenever the buffer changes.

The ideal fix would be to somehow launch servers before any blocking
requests, but that doesn't seem trivial.  So for now, let's add an
error message that informs the user that the *-sync commands are
currently not supported without running lsp-enable first.

Came up in #582

This adds a new option that is set to "fail" whenever LSP is not
enabled, so:
Closes #461