Work around inconsistent wcwidth by using ASCII instead of emoji
As reported on Discord, macOS ships with an libc where wcwidth()
returns 1 for emoji characters, contradicting modern Unicode which
specifies width 2.
Some applications work around this issue by bundling their
own (more modern) wcwidth(). This inconsistency makes things yet more
problematic: Kakoune thinks the width of an emoji character is 1,
but the terminal might use 2. This causes a rendering glitch.

Work around this by substituting the emoji with ASCII chars if the
system wcwidth() returns != 2 for emoji.  See this stackoverflow page
for a discussion on how to do that:
https://unix.stackexchange.com/questions/245013/get-the-display-width-of-a-string-of-characters

Closes #544