Fix regression when applying text edits at buffer end
The new edit coalescing logic fails to handle trailing (and possibly
leading) blank lines. For example, lsp-formatting-sync on a Go
file with a trailing blank line sends a text edit which deletes the
blank line:

    {
      "range": {
        "start": {
          "line": 31,
          "character": 0
        },
        "end": {
          "line": 32,
          "character": 0
        }
      },
      "newText": ""
    }

Line 32 is empty, which means we fail to compute the byte offset of
the character. Let's default to zero in that case.