Fix buffer-overflow when applying text edits until buffer end
This is another fallout from the new text-edit-coalescing logic.
Commit 8c0b2fc (Fix regression when applying text edits at buffer end,
2022-01-26) fixed handling text edits that range up to (including)
an empty line, but there is also the case where the end of the range
is a line that doesn't exist. This means that the text edit replaces
everything until the buffer end.

In this case we can't even access the line, so let's not panic.

Fixes #589